requirements.txt
pyTelegramBotAPI==4.23.0
psycopg2-binary==2.9.10

psutil~=6.1.0